+++ packages/components/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 2.13.2 (2026-03-19)
+
+**Note:** Version bump only for package @ttoss/components
+
 ## [2.13.1](https://github.com/ttoss/ttoss/compare/@ttoss/components@2.13.0...@ttoss/components@2.13.1) (2026-02-23)
 
 **Note:** Version bump only for package @ttoss/components