+++ packages/forms/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.42.0](https://github.com/ttoss/ttoss/compare/@ttoss/forms@0.41.2...@ttoss/forms@0.42.0) (2026-03-19)
+
+### Features
+
+- **forms:** generic FormFieldPhone with configurable and editable country code ([#892](https://github.com/ttoss/ttoss/issues/892)) ([0ef9c71](https://github.com/ttoss/ttoss/commit/0ef9c71efbf902ce94312068c5995dcfdb6df4e4))
+
 ## 0.41.2 (2026-03-19)
 
 **Note:** Version bump only for package @ttoss/forms